requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.6
Flask==3.1.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
MarkupSafe==3.0.2
multidict==6.1.0
openai==1.57.3
propcache==0.2.1
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.19
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tqdm==4.67.1
twilio==9.4.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
Werkzeug==3.1.3
yarl==1.18.3
